Upload files to Canvas

Upload a local file to a location on Canvas LMS. Read more about possible parameters at https://canvas.instructure.com/doc/api/file.file_uploads.html.

cnvs_upload(endpoint, path, ...)

Arguments

endpoint
Character string of the API endpoint for upload. This depends on the required context of file. For example, uploading a file to a course uses the endpoint "/api/v1/courses/:course_id/files". Uploading a file as an assignment submission has the endpoint "/api/v1/courses/:course_id/assignments/:assignment_id/submissions/:user_id/files"

path
Path to file to upload

...
Other parameters passed along to cnvs(). Use these to specify endpoint parameters, e.g. course_id, parameters for the file upload, e.g. parent_folder_path, or other parameters to cnvs() like .token or .api_endpoint.

Examples

if (FALSE) {
## Upload a course file, ends up in "unfiled"
cnvs_upload("/api/v1/courses/:course_id/files",
  path = "notes.pdf", course_id = "1732420")

## Upload a course file to a specific folder
cnvs_upload("POST /api/v1/courses/:course_id/files",
  path = "notes.pdf", course_id = "1732420", parent_folder_path = "cnvs_files")
}